+++ py-scouter/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod agent;
 pub mod alert;
+pub mod bifrost;
 pub mod client;
-pub mod dataset;
 pub mod drift;
 pub mod evaluate;
 pub mod logging;
@@ -17,7 +17,7 @@ use pyo3::prelude::*;
 
 #[pymodule]
 fn _scouter(m: &Bound<'_, PyModule>) -> PyResult<()> {
-    dataset::add_dataset_module(m)?;
+    bifrost::add_dataset_module(m)?;
     queue::add_queue_module(m)?;
     logging::add_logging_module(m)?;
     client::add_client_module(m)?;
